Refactored index and cwd
